Refactor ctrl server and client (#4234)
* refactor CtrlClient

* RpcServer

* use Rpc

* Update rpc_client.cpp

add comment
